requirements.txt
Flask
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-jaeger-thrift
opentelemetry-instrumentation-flask